Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #iclude<stdio.h>
 int main(){
-    float num;
-    scanf("%f",&num);
+    int num;
+    scanf("%d",&num);
     printf("Area: %f",num*3.14);
     return 0;
 }
